requirements.txt
python-dotenv
pandas
numpy
pymango
certifi
pymongo[srv]==3.12
scikit-learn
#-e .